fix: unblock stg_order_down+ models, cast NUMERIC to FLOAT64 for XGBoost compat
mart_ml_features/mart_ptf_lag_features were stuck excluded pending an
order_down backfill that turned out to be a table-wide duplication +
nanosecond-timestamp issue (now fully deduped), so remove stg_order_down
and its 7 downstream models from DBT_EXCLUDE_PENDING_BACKFILL.

Also cast NUMERIC columns to FLOAT64 in stg_imbalance/stg_idm_transactions
staging models — BigQuery NUMERIC comes back as pandas Decimal/object,
which was breaking ptf_hourly_inference's XGBoost predict() call on
net_imbalance_mwh.

+++ epias_dbt/models/staging/stg_idm_transactions.sql
@@ -19,8 +19,8 @@ SELECT
     CAST(date AS DATE) AS date,
     CAST(SUBSTR(CAST(hour AS STRING), 1, 2) AS INT64) AS hour,
     CAST(contractName AS STRING) AS contract_name,
-    CAST(price    AS NUMERIC) AS price_try,
-    CAST(quantity AS NUMERIC) AS quantity_mwh,
+    CAST(price    AS FLOAT64) AS price_try,
+    CAST(quantity AS FLOAT64) AS quantity_mwh,
 
     {% if 'buyerOrganizationId' in silver_cols %}
     CAST(buyerOrganizationId  AS INT64) AS buyer_organization_id,
